pyproject.toml
[project]
name = "hunting-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Shukurullo-98",email = "tbekamov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
